Update README.md
This Arduino program controls an LED using a push button. The LED is connected to pin 13 and the button to pin 7. In setup(), pin 13 is set as OUTPUT and pin 7 as INPUT. In loop(), the button state is read. If pressed (HIGH), the LED turns ON; if not pressed (LOW), the LED turns OFF. It shows simple digital input-output control.